feat: extract toolbar collapse observer
